+++ src/constants.js
@@ -1,0 +1,13 @@
+export const PlatformName = "Meraki";
+export const PluginName = "homebridge-meraki-control";
+
+export const ApiUrls = {
+    "Base": "/api/v1",
+    "OrganizationNetworks": "/organizations/organizationId/networks",
+    "OrganizationDevices": "/organizations/organizationId/devices",
+    "NetworkId": "/networks/networkId",
+    "AppliancePorts": "/networks/networkId/appliance/ports",
+    "MrSsids": "/networks/networkId/wireless/ssids",
+    "MsPorts": "/devices/serialNumber/switch/ports",
+    "DbClients": "/networks/networkId/clients"
+}